7 miglioramenti

+++ src/Win7POS.Core/Import/ImportDiffItem.cs
@@ -4,9 +4,9 @@ namespace Win7POS.Core.Import
     {
         public string Barcode { get; set; } = string.Empty;
         public string IncomingName { get; set; } = string.Empty;
-        public int IncomingPrice { get; set; }
+        public long IncomingPrice { get; set; }
         public string ExistingName { get; set; } = string.Empty;
-        public int? ExistingPrice { get; set; }
+        public long? ExistingPrice { get; set; }
         public ImportDiffKind Kind { get; set; }
     }
 }
