+++ src/Win7POS.Core/Import/ImportRow.cs
@@ -4,7 +4,7 @@ namespace Win7POS.Core.Import
     {
         public string Barcode { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
-        public int UnitPrice { get; set; }
+        public long UnitPrice { get; set; }
         public int? Cost { get; set; }
         public int? Stock { get; set; }
     }
