+++ src/Win7POS.Core/Reports/DailyTakings.cs
@@ -3,9 +3,9 @@ namespace Win7POS.Core.Reports
     public sealed class DailyTakings
     {
         public int TotalSalesCount { get; set; }
-        public int GrossTotal { get; set; }
-        public int CashTotal { get; set; }
-        public int CardTotal { get; set; }
-        public int ChangeTotal { get; set; }
+        public long GrossTotal { get; set; }
+        public long CashTotal { get; set; }
+        public long CardTotal { get; set; }
+        public long ChangeTotal { get; set; }
     }
 }
